Processing configuration after load
When I was working to incorporate this into a Rails application the
configuration options were throwing lots of errors related to having
files loaded. The sequence of initializers had become important and
labrythine.

To solve this I am instead capturing the configuration block and
executing it later in the Rails load sequence. There may be some
interesting development behaviors that could occur.

From
http://guides.rubyonrails.org/configuring.html#initialization-events
> to_prepare: Run after the initializers are run for all Railties
> (including the application itself), but before eager loading and
> the middleware stack is built. More importantly, will run upon every
> request in development, but only once (during boot-up) in production
> and test.